nixd/fork: rename ServerSerialization -> JSONSerialization
