+++ lib/nixd/include/nixd/Server.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "nixd/EvalDraftStore.h"
+#include "nixd/JSONSerialization.h"
 
 #include "lspserver/Connection.h"
 #include "lspserver/DraftStore.h"
@@ -8,61 +9,15 @@
 #include "lspserver/LSPServer.h"
 #include "lspserver/Logger.h"
 #include "lspserver/Path.h"
-#include "lspserver/Protocol.h"
 #include "lspserver/SourceCode.h"
 
 #include <llvm/ADT/FunctionExtras.h>
-#include <llvm/Support/JSON.h>
 #include <llvm/Support/raw_ostream.h>
 
 #include <cstdint>
 #include <memory>
 
-// Extension to `lspserver`
-namespace lspserver {
-
-llvm::json::Value toJSON(const CompletionContext &R);
-
-llvm::json::Value toJSON(const TextDocumentPositionParams &R);
-
-llvm::json::Value toJSON(const CompletionParams &R);
-
-bool fromJSON(const llvm::json::Value &Params, CompletionItem &R,
-              llvm::json::Path P);
-
-bool fromJSON(const llvm::json::Value &Params, CompletionList &R,
-              llvm::json::Path P);
-
-bool fromJSON(const llvm::json::Value &Params, Hover &R, llvm::json::Path P);
-
-bool fromJSON(const llvm::json::Value &Params, MarkupContent &R,
-              llvm::json::Path P);
-
-} // namespace lspserver
-
 namespace nixd {
-
-namespace configuration {
-
-struct InstallableConfigurationItem {
-  std::vector<std::string> args;
-  std::string installable;
-};
-
-struct TopLevel {
-  /// Nix installables that will be used for root translation unit.
-  std::optional<InstallableConfigurationItem> installable;
-
-  /// Get installable arguments specified in this config, fallback to file \p
-  /// Fallback if 'installable' is not set.
-  [[nodiscard]] std::tuple<nix::Strings, std::string>
-  getInstallable(std::string Fallback) const;
-};
-
-bool fromJSON(const llvm::json::Value &Params, TopLevel &R, llvm::json::Path P);
-bool fromJSON(const llvm::json::Value &Params, InstallableConfigurationItem &R,
-              llvm::json::Path P);
-} // namespace configuration
 
 struct CompletionHelper {
   using Items = std::vector<lspserver::CompletionItem>;
@@ -75,38 +30,10 @@ struct CompletionHelper {
                              const nix::StaticEnv &SEnv);
 };
 
-using WorkspaceVersionTy = uint64_t;
-
-namespace ipc {
-
-/// Messages sent by workers must tell it's version
-struct WorkerMessage {
-  WorkspaceVersionTy WorkspaceVersion;
-};
-
-bool fromJSON(const llvm::json::Value &, WorkerMessage &, llvm::json::Path);
-llvm::json::Value toJSON(const WorkerMessage &);
-
-/// Sent by the worker process, tell us it has prepared diagnostics
-/// <----
-struct Diagnostics : WorkerMessage {
-  std::vector<lspserver::PublishDiagnosticsParams> Params;
-};
-
-bool fromJSON(const llvm::json::Value &, lspserver::PublishDiagnosticsParams &,
-              llvm::json::Path);
-
-bool fromJSON(const llvm::json::Value &, Diagnostics &, llvm::json::Path);
-llvm::json::Value toJSON(const Diagnostics &);
-
-struct Completion : WorkerMessage {
-  lspserver::CompletionList List;
-};
-
-} // namespace ipc
-
 /// The server instance, nix-related language features goes here
 class Server : public lspserver::LSPServer {
+
+  using WorkspaceVersionTy = ipc::WorkspaceVersionTy;
 
   int WaitWorker = 0;
